slight adjustment on thinking time

+++ common_core/philosophers/philo/philo.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   philo.h                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: cseriildii <cseriildii@student.42.fr>      +#+  +:+       +#+        */
+/*   By: icseri <icseri@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/24 13:20:04 by icseri            #+#    #+#             */
-/*   Updated: 2024/08/08 18:31:46 by cseriildii       ###   ########.fr       */
+/*   Updated: 2024/08/09 13:00:05 by icseri           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -50,7 +50,7 @@ typedef struct s_philo
 {
 	int				id;
 	int				type;
-	long			initial_thinking_time;
+	int				initial_thinking_time;
 	int				status;
 	int				times_eaten;
 	long long		last_eating_time;
